python-flask-spid/requirements.txt
flask==1.0
pyOpenSSL==18.0.0
python-saml==2.4.2
decorator==4.3.0
nose
rfc5424-logging-handler
swagger-ui-bundle==0.0.2
connexion == 2.2.0
PyJWT